% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d07_computingApproxModifDuration.R
\name{approxModifDuration}
\alias{approxModifDuration}
\title{Calculates the Approximate Modified Duration.}
\usage{
approxModifDuration(pvBase, pvPlus, pvMinus, percentChangeYtm)
}
\arguments{
\item{pvBase}{A number.}

\item{pvPlus}{A number.}

\item{pvMinus}{A number.}

\item{percentChangeYtm}{A number.}
}
\value{
Input values to four arguments  \code{pvBase},\code{pvPlus} ,\code{pvMinus}  and \code{percentChangeYtm}.
}
\description{
Calculates the Approximate Modified Duration.
}
\details{
According to information provided by Adams and Smith (2019), the method \code{approxModifDuration()} is developed to calculate the Approximate Modified Duration.
}
\examples{
approxModifDuration(pvBase=99.956780,pvPlus=100.631781,pvMinus=101.250227,percentChangeYtm=0.0005)
approxModifDuration (pvBase=42.223649,pvPlus=42.100694,pvMinus=42.346969,percentChangeYtm=0.0001)
}
\references{
Adams,J.F. & Smith,D.J.(2019). Understanding Fixed‑Income Risk and Return. In CFA Program Curriculum 2020 Level I Volumes 1-6. (Vol. 5, pp. 237-299). Wiley Professional Development (P&T). ISBN 9781119593577, \url{https://bookshelf.vitalsource.com/books/9781119593577}
}
\author{
MaheshP Kumar, \email{maheshparamjitkumar@gmail.com}
}
