% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/38_bondPricingUsingForwardRates.R
\name{frPricing}
\alias{frPricing}
\title{Calculates Bond Price using the Forward Rate Input.}
\usage{
frPricing(cpns, fri, mv, n)
}
\arguments{
\item{cpns}{A vector.}

\item{fri}{A vector.}

\item{mv}{A number.}

\item{n}{A number.}
}
\value{
Input values to four arguments  \code{cpns} ,\code{fri}, \code{mv} and \code{n}.
}
\description{
Calculates Bond Price using the Forward Rate Input.
}
\details{
According to information provided by Adams and Smith (2019), the method \code{frPricing()} is developed to compute the Bond Price using the Forward Rate Input. If 0y1y is 1.88 percent, 1y1y is 2.77 percent and 2y1y is 3.54 percent then the bond price works out to be 102.965 dollars. Here, \code{cpns} is vector of coupon Payments for three years,\code{fri} is a vector of that contains three values that serve as forward rate input factor,\code{mv}is maturity Value, and \code{n} is number of years that is 3 years.
}
\examples{
frPricing(cpns=c(3.75,3.75,3.75),fri=c((1.0188),(1.0188*1.0277),(1.0188*1.0277*1.0354)),mv=100,n=3)
}
\references{
Adams,J.F. & Smith,D.J.(2019). Introduction to fixed-income valuation. In CFA Program Curriculum 2020 Level I Volumes 1-6. (Vol. 5, pp. 107-151). Wiley Professional Development (P&T). ISBN 9781119593577, \url{https://bookshelf.vitalsource.com/books/9781119593577}
}
\author{
MaheshP Kumar, \email{maheshparamjitkumar@gmail.com}
}
