% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d10_computingMoneyDuration.R
\name{moneyDuration}
\alias{moneyDuration}
\title{Calculates Money Duration of a Bond.}
\usage{
moneyDuration(macDuration, ytm, pvFullBondPrice)
}
\arguments{
\item{macDuration}{A number.}

\item{ytm}{A number.}

\item{pvFullBondPrice}{A number.}
}
\value{
Input values to three arguments  \code{macDuration},\code{ytm} and \code{pvFullBondPrice}.
}
\description{
Calculates Money Duration of a Bond.
}
\details{
According to information provided by Adams and Smith (2019), the method \code{moneyDuration()} is developed to calculate Money Duration of a Bond.
}
\examples{
moneyDuration(macDuration=7.0029,ytm=0.104,pvFullBondPrice=85.5031)
}
\references{
Adams,J.F. & Smith,D.J.(2019). Understanding Fixed‑Income Risk and Return. In CFA Program Curriculum 2020 Level I Volumes 1-6. (Vol. 5, pp. 237-299). Wiley Professional Development (P&T). ISBN 9781119593577, \url{https://bookshelf.vitalsource.com/books/9781119593577}
}
\author{
MaheshP Kumar, \email{maheshparamjitkumar@gmail.com}
}
